MSRP: harden per-hop latency accumulation against overflow
Two overflow bugs the high-effort review found in the latency path:
  * srpPortTcMaxLatency computed `maxFrameSize * 8_000_000` in UInt, which is
    1.6e10 for the default 2000-byte frame -- fine on arm64 (64-bit UInt) but a
    trap on the armhf build (32-bit UInt). Compute the frame time in UInt64.
  * _applyStreamPlan added this hop's latency to the peer-supplied
    accumulatedLatency with `+=` on a UInt32, so a received value near UInt32.max
    overflowed and trapped -- a remotely triggerable DoS on the recompute path.
    Saturate at UInt32.max instead.

Tests: the existing srpPortTcMaxLatency test already exercises the overflowing
intermediate (traps on armhf pre-fix); add a relay test asserting a near-max
received latency saturates rather than wraps.

+++ Sources/MRP/Applications/MSRP/MSRPTypes.swift
@@ -433,8 +433,10 @@ func srpPortTcMaxLatency(
 ) -> Int {
   var latency = meanLinkDelayNs
   if linkSpeedKbps > 0 {
-    // one max-size frame time in ns: bytes * 8 bits / (kbps * 1000) s, scaled to ns
-    let frameTimeNs = Int(maxFrameSize * 8_000_000 / linkSpeedKbps)
+    // one max-size frame time in ns: bytes * 8 bits / (kbps * 1000) s, scaled to ns.
+    // Compute in UInt64: maxFrameSize * 8_000_000 (e.g. 2000 * 8e6 = 1.6e10) overflows a 32-bit
+    // UInt and traps on the armhf target before the divide.
+    let frameTimeNs = Int(UInt64(maxFrameSize) * 8_000_000 / UInt64(linkSpeedKbps))
     latency += 2 * frameTimeNs // c) store-and-forward + b) one interfering lower-priority frame
   }
   return latency
